Increase the technical point limits on write
Technically, 8191 points can be saved into a XY record. The limit is imposed by the size field of the record.
Previously gdsCAD used a signed integer for the length field, which halves the available number of points.

The point limit can be circumvented by Boarders by using multiple XY entries. Not all GDS readers support this
unfortunately, so a warning is emitted in verbose mode.