Fix GDS import function for nested Cells and large designs
In case of nested Cell depenencies, we cannot add a Cell which
may refer to cells where the dependencies are not yet solved.

So first resolve all depenencies and then add them to the library.

Furthermore, gdscad has the concept of units and precicion,
whereas is not represented in the GDS file format. So care has to
be taken on how to handle this.